pyproject.toml
[tool.poetry]
name = "monitor"
version = "0.1.1"
description = "Small conceptual website monitor using Kafka and PostgreSQL."
authors = ["Sören König <soeren.koenig@mailbox.org>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/skoenig/website-monitor"
packages = [{ include = "monitor" }]

[tool.poetry.dependencies]
python = "3.13.*"
PyYAML = "6.0.2"
apscheduler = "^3.6.3"
requests = "^2.32"
kafka-python = "2.2.14"
psycopg2-binary = "2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.1"
coverage = "^7.0.0"
pre-commit = "*"
ruff = "^0.12.0"

[tool.ruff]
lint.ignore = ["E712"]
lint.per-file-ignores = {"*.ipynb" = ["ALL"]}
format.quote-style = "single"

[tool.yamlfix]
sequence_style = "block_style"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
